Update the display date upon liking, fix a bug like text bug
(Like text was looking for the wrong class name)
The date is tricky part here, right? We're updating the date based on the server. The user will their local date when they click update. The only real solution to this would be to use UTC for the date and then adjust for timezone offset in the browser.